now program check the directory if not ound it creates one and put that file in to that

+++ main.js
@@ -1,22 +1,9 @@
 const electron = require('electron');
-const os = require('os');
 const fs = require('fs');
 const path = require('path');
 
 // use ipcRenderer to access All node and electron methods in html file
 const {ipcRenderer} = electron;
-
-// create ul, li and assign cpus function to normal variable
-const ul = document.getElementById('list');
-let cpus = os.cpus();
-
-// simply get All cpus and iterate on it
-cpus.forEach(cpu => {
-    const li = document.createElement('li');
-    let content = document.createTextNode(`CPU MODEL : ${cpu.model} CPU SPEED : ${cpu.speed} \n`);
-    li.appendChild(content);
-    ul.appendChild(li);
-});
 
 // create all variables
 fileName = document.getElementById('fileName');
@@ -26,10 +13,24 @@ pathName = path.join(__dirname, 'Files');
 
 createFile.addEventListener('click', (e) => {
     e.preventDefault();
+    // check to see if file exist in directory or not
+    if(!fs.existsSync(pathName)) {
+        fs.mkdirSync(pathName);
+        createFileForMe();
+    } else {
+        createFileForMe();
+    }
+});
+
+var createFileForMe = () => {
     let file = path.join(pathName, fileName.value);
     let content = fileContent.value;
-    fs.writeFile(file, content, (err) => {
-        if (err) throw err;
-        console.log('file is created');
-    });
-});
+    if(fs.existsSync(file)) {
+        console.log('file is exist');
+    } else {
+        fs.writeFile(file, content, (err) => {
+            if (err) throw err;
+            console.log('file is created');
+        });
+    }
+}
